==> research/05_model_evaluation.ipynb <==
import os
os.chdir("../")
%pwd

os.environ["MLFLOW_TRACKING_URI"]="https://dagshub.com/Stanlito-AI/MLOps_ML.mlflow"
os.environ["MLFLOW_TRACKING_USERNAME"]="Stanlito-AI"
os.environ["MLFLOW_TRACKING_PASSWORD"]="f616a6150e177ffd52787126e7241eeb312f247f"


from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class ModelEvaluationConfig:
    root_dir: Path
    test_data_X: Path
    test_data_y: Path
    model_path: Path
    all_params: dict  
    metric_file_name: Path
    target_column: str
    mlflow_uri: str

from mlhrtds.constants import *
from mlhrtds.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])


    def get_model_evaluation_config(self) -> ModelEvaluationConfig:
        config = self.config.model_evaluation
        params = self.params.CatBoostRegressor
        schema =  self.schema.TARGET_COLUMN

        create_directories([config.root_dir])

        model_evaluation_config = ModelEvaluationConfig(
            root_dir=config.root_dir,
            test_data_X=config.test_data_X,
            test_data_y=config.test_data_y,
            model_path = config.model_path,
            all_params=params,
            metric_file_name = config.metric_file_name,
            target_column = schema.name,
            mlflow_uri="https://dagshub.com/Stanlito-AI/MLOps_ML.mlflow",
           
        )

        return model_evaluation_config

import os
import pandas as pd
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
from urllib.parse import urlparse
import mlflow
import mlflow.sklearn
import numpy as np
import joblib
# from mlds.entity.config_entity import ModelEvaluationConfig
from mlhrtds.utils.common import save_json
from pathlib import Path


class ModelEvaluation:
    def __init__(self, config: ModelEvaluationConfig):
        self.config = config

    
    def eval_metrics(self,actual, pred):
        rmse = np.sqrt(mean_squared_error(actual, pred))
        mae = mean_absolute_error(actual, pred)
        r2 = r2_score(actual, pred)
        return rmse, mae, r2
    


    def log_into_mlflow(self):

        test_x = pd.read_csv(self.config.test_data_X)
        test_y = pd.read_csv(self.config.test_data_y)
        model = joblib.load(self.config.model_path)

        # test_x = test_data.drop([self.config.target_column], axis=1)
        # test_y = test_data[[self.config.target_column]]


        mlflow.set_registry_uri(self.config.mlflow_uri)
        tracking_url_type_store = urlparse(mlflow.get_tracking_uri()).scheme


        with mlflow.start_run():

            predicted_qualities = model.predict(test_x)

            (rmse, mae, r2) = self.eval_metrics(test_y, predicted_qualities)
            
            # Saving metrics as local
            scores = {"rmse": rmse, "mae": mae, "r2": r2}
            save_json(path=Path(self.config.metric_file_name), data=scores)

            mlflow.log_params(self.config.all_params)

            mlflow.log_metric("rmse", rmse)
            mlflow.log_metric("r2", r2)
            mlflow.log_metric("mae", mae)


            # Model registry does not work with file store
            if tracking_url_type_store != "file":

                # Register the model
                # There are other ways to use the Model Registry, which depends on the use case,
                # please refer to the doc for more information:
                # https://mlflow.org/docs/latest/model-registry.html#api-workflow
                mlflow.sklearn.log_model(model, "model", registered_model_name="CatBoostRegressor")
            else:
                mlflow.sklearn.log_model(model, "model")

    

try:
    config = ConfigurationManager()
    model_evl_config = config.get_model_evaluation_config()
    model_evl = ModelEvaluation(config=model_evl_config)
    model_evl.log_into_mlflow()
except Exception as e:
    raise e
